Cell-cell communication from single cell data

Calculate condition-specific cell-cell communication mediated by secreted proteins from scRNA-Seq data.

Usage

SecAct.CCC.scRNAseq(
  Seurat_obj,
  cellType_meta,
  condition_meta,
  conditionCase,
  conditionControl,
  scale.factor = 1e+05,
  act_diff_cutoff = 2,
  exp_logFC_cutoff = 0.2,
  exp_mean_all_cutoff = 2,
  exp_fraction_case_cutoff = 0.1,
  padj_cutoff = 0.01,
  sigMatrix = "SecAct",
  is.group.sig = TRUE,
  is.group.cor = 0.9,
  lambda = 5e+05,
  nrand = 1000
)

Arguments

Seurat_obj

A Seurat object.

cellType_meta

Column name in meta data that includes cell-type annotations.

condition_meta

Column name in meta data that includes condition information.

conditionCase

Case condition.

conditionControl

Control condition.

scale.factor

Sets the scale factor for cell-level normalization in step2.

act_diff_cutoff

Cut off for activity change (i.e., z score) in step 1.

exp_logFC_cutoff

Cut off for log fold change in step 2.

exp_fraction_case_cutoff

Cut off for the fraction of cells expressing secreted protein-coding genes in step 2.

padj_cutoff

Adjusted p value cut off.

sigMatrix

Secreted protein signature matrix.

is.group.sig

A logical indicating whether to group similar signatures.

is.group.cor

Correlation cutoff of similar signatures.

lambda

Penalty factor in the ridge regression.

nrand

Number of randomization in the permutation test, with a default value 1000.

Value

A Seurat object.