Cell state activity inference from single cell data

Calculate secreted protein signaling activity of cell states from single cell RNA-Sequencing data.

Usage

SecAct.activity.inference.scRNAseq(
  inputProfile,
  cellType_meta,
  is.singleCellLevel = FALSE,
  sigMatrix = "SecAct",
  is.filter.sig = FALSE,
  is.group.sig = TRUE,
  is.group.cor = 0.9,
  lambda = 5e+05,
  nrand = 1000
)

Arguments

inputProfile

A Seurat object.

cellType_meta

Column name in meta data that includes cell-type annotations.

is.singleCellLevel

A logical flag indicating whether to calculate for each single cell (Default: FALSE).

sigMatrix

Secreted protein signature matrix.

is.filter.sig

A logical flag indicating whether to filter the secreted protein signatures based on the genes from inputProfile (Default: FALSE). Because some sequencing platforms (e.g., CosMx) cover only a subset of secreted proteins, setting this option to TRUE restricts the activity inference on those proteins.

is.group.sig

A logical flag indicating whether to group similar signatures (Default: TRUE). Many secreted proteins, such as cytokines with similar cell surface receptors and downstream pathways, have cellular effects that appear redundant within a cellular context. When enabled, this option clusters secreted proteins based on Pearson correlations among their composite signatures. The output still reports activity estimates for all secreted proteins prior to clustering. Secreted proteins assigned to the same non-redundant cluster share the same inferred activity.

is.group.cor

A numeric value specifying the correlation cutoff used to define similar signatures (Default: 0.90). When r > 0.90, 1,170 secreted protein signatures are grouped into 657 non-redundant signature groups.

lambda

Penalty factor in the ridge regression.

nrand

Number of randomization in the permutation test, with a default value 1000.

Value

A Seurat object.