Secreted protein signaling velocity

Calculate the signaling velocity of secreted proteins based on their activities.

Usage

SecAct.signaling.velocity.scST(
  SpaCET_obj,
  sender,
  secretedProtein,
  receiver,
  cellType_meta,
  scale.factor = 1e+05,
  CustomizedAreaCoordinates = NULL,
  radius = 20
)

Arguments

SpaCET_obj

A SpaCET object.

sender

Sender cell types.

secretedProtein

Secreted proteins.

receiver

Receiver cell types.

scale.factor

Sets the scale factor for spot-level normalization.

radius

Radius cut off (unit: um).

Value

A ggplot2 object.

Details

The velocity direction starts from the source cell producing a secreted protein and moves to sink cells receiving the secreted protein signal. The velocity magnitude represents the product between the secreted protein-coding gene expression at source cells and signaling activities at sink cells.

Examples

SecAct.signaling.velocity.scST(SpaCET_obj, sender="Fibroblast", secretedProtein="THBS2", receiver="Tumor_boundary", cellType_meta="cellType")
SecAct.signaling.velocity.scST(SpaCET_obj, sender="Fibroblast", secretedProtein="THBS2", receiver="Tumor_boundary", cellType_meta="cellType", CustomizedAreaCoordinates=c(8290,8366,1100,1400))